Fix name conflict with FileSystemType from winnt.h
winnt.h has this enum in global scope:
  typedef enum _CM_SERVICE_NODE_TYPE {
    ...
    FileSystemType = SERVICE_FILE_SYSTEM_DRIVER,
    ...
  } SERVICE_NODE_TYPE;

Clang complains that references to FileSystemType are ambiguous.  It
suggests using 'enum FileSystemType' to resolve the ambiguity, but that
seems weird in C++.

Instead, I changed the uses in this file to fileapi::FileSystemType and
left a comment where the using decl would naturally go.

R=kinuko@chromium.org
BUG=82385

Review URL: https://codereview.chromium.org/14615015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@200813 4ff67af0-8c30-449e-8e8b-ad334ec8d88c