Restore functionality of chrome.location API. CL (https://codereview.chromium.org/15000006/) disables functionality of chrome.location API is is intended for merging to M28.
The original reason for disabling it is likely resolved by already landed CLs:
https://codereview.chromium.org/14869009/
https://codereview.chromium.org/14969010/

The only way to find out whether the above fixes were successful is to enable the API again (in Canary/Dev and for component extensions). This is what this CL is doing. It won't be merged into M28.

BUG=230943

Review URL: https://chromiumcodereview.appspot.com/14788017

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@201019 4ff67af0-8c30-449e-8e8b-ad334ec8d88c