Fix crash on resuming Drive file upload.
This is rather a workaround, than a fundamental fix that should
come in a week or so (the true fix is to kill OperationRegistry completely:
crbug.com/164098).

GetUploadStatusOperation is called in DriveUploader::ResumeUpload,
which is a "resume" from user point of view, but is to "start" one new series
of upload sequence from lower-level point of view.
Thus, it has to initiate with NotifyStart() rather than NotifyResume().

In the test code, GetUploadStatusOperation is also used as an
intermediate operation during a series of upload operations.
Under this patch such tests will not work. As a temporal solution,
this patch disables the tests as well. I hope they'll be back very soon,

BUG=242141

Review URL: https://chromiumcodereview.appspot.com/15444002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@201073 4ff67af0-8c30-449e-8e8b-ad334ec8d88c