Instant Extended: Hide info bar tip when instant overlay is visible
This CL hides the info bar tip when showing an instant overlay. This
prevents the infobar anti-spoof arrow from being visible under the
overlay.

BUG=174077
TEST=Ran chrome with instant extended enabled. Navigated to
  http://www.consulfrance-toronto.org/
and waited for the translate info bar. Typed http://cnn.com in the
omnibox. Verified that the infobar arrow was not visible.

Review URL: https://chromiumcodereview.appspot.com/14905009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@201196 4ff67af0-8c30-449e-8e8b-ad334ec8d88c