Introduce a priority queue to the image loader.
Image loader used to start loading all requests at once, which caused long loading times in some scenarios. Eg. while images in the mosaic view are being loaded in the background, switching to the next image in the slide view is slow. It may happen that there are hundreds of pending images (for mosaic view preload) which have to be handled before the one is important (slide view).

To fix this issue (1) priorities were introduces, (2) tasks are queued, (3) number of tasks performed in parallel is limited to 5.
In practice, it gives a guarantee that there will be at most 5 tasks with lower priority executed before a task with higher priority. Taking into account an average resizing time, which is around 200ms, this gives a guarantee that on average, the image with the highest priority will be handled in the worst case with 1 second of delay.

Along the way, the selected images in the mosaic view have higher priority than other tiles. Also, background thumbnail generation for copying images has now very low priority to minimize load of important images.

TEST=Tested manually. 
BUG=239237

Review URL: https://chromiumcodereview.appspot.com/14623021

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@201245 4ff67af0-8c30-449e-8e8b-ad334ec8d88c