Fix the problem that black background window might occasionally appear in front on Windows
When a new panel is added to a stack, the following things occur:
1) The panel is initially created as always on top on Windows.
2) The panel is changed to non-always-on-top.
3) The background stack window is made to the owner of the new panel window.
4) The background stack window is set to behind the new panel window in z-order.

Occasionally, step 2 might be run after step 4. When this occurs, after step 1, 3 and 4, the background stack window will automatically obtain the always-on-top state since the panel window behind which is set to is always-on-top. For most of time, running step 2 after 4 will remove the always-on-top state for both new panel window and background stack window. However, if chrome is not foreground application, Windows system might fail to deprive the always-on-top state for the background stack window.

The fix is to create the panel window with always-on-top state depending on its collection.

BUG=241224
TEST=Manual test by following the bug report

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=200788

R=dimich@chromium.org

Review URL: https://codereview.chromium.org/14956006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@201325 4ff67af0-8c30-449e-8e8b-ad334ec8d88c