Add flag for drawing layers to screen with Ganesh
The --force-direct-layer-drawing flag causes any layer that could be
drawn direct to the backbuffer to be drawn to the backbuffer using
Ganesh.  This flag also requires both the --enable-threaded-compositing
flag and the --enable-impl-side-painting flag to have any effect.

This patch also turns on testing for PictureDrawQuad using Ganesh in the
cc pixeltests.

BUG=none

Review URL: https://chromiumcodereview.appspot.com/13863015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@201362 4ff67af0-8c30-449e-8e8b-ad334ec8d88c