sync: Test and improve bookmark performance
This commit adds a test for bookmark model association.  It's not a
great test in terms of code coverage, but it's possible to turn it into
a useful benchmark by changing the kNumBookmarksPerFolder and
kNumFolders constants.

Also included in this commit is a small change to the WriteTransaction's
SaveOriginal method.  The old code created a temporary
EntryKernelMutation, which had to be cleaned up as it went out of scope.
It turns out that destroying the four EntitySpecifics objects it
contains can be pretty expensive.

The performance impact isn't huge, but it is visible in the profiling
results.  Since the fix is so trivial, the cost-benefit tradeoff is
pretty clearly in favor of fixing this.

BUG=241813,238621

Review URL: https://chromiumcodereview.appspot.com/15308003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@201363 4ff67af0-8c30-449e-8e8b-ad334ec8d88c