Fixed crasher initializing the AppPackUpdater.
The browser is restarted with the --login-user switch when it crashes. In retail
mode, this causes the creation of the signin Profile after the "normal", retail
mode Profile has been created; that causes the AppPackExternalLoader to be
requested twice, resulting in a NULL pointer access.

This fix checks if the AppPackExternalLoader is NULL, so that it also works
for multi-profiles in retail mode.

BUG=233414

Review URL: https://chromiumcodereview.appspot.com/15484008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@201376 4ff67af0-8c30-449e-8e8b-ad334ec8d88c