The chrome browser target should not depend on plugins.
This in turn brings with it a dependency on webkit_glue which in turn brings in webkit.
We should only depend on plugins_common in the browser. 

The pepper component updater for flash is the only component in the browser which depends
on webkit\plugins\ppapi. This is to check if the interface name passed in is supported by the browser.

I added a function IsSupportedPepperInterface to the newly added ppapi_utils.cc/.h file which lives in
webkit/common/plugins/ppapi

Added a macro LEGACY_IFACE to the newly added ppapi/thunk/interfaces_legacy.h file. This macro consolidates
the list of interfaces being checked for on the browser. It is used in plugin_module.cc and ppapi_utils.cc

BUG=237249
TBR=cpu@chromium.org, jamesr@chromium.org, raymes@chromium.org, yzshen@chromium.org

Review URL: https://codereview.chromium.org/15505004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@201623 4ff67af0-8c30-449e-8e8b-ad334ec8d88c