Translate: infobars should not appear when a page has a refresh meta tag
If a page contains a <meta http-equiv="refresh" content="0;...">,
a client redirect is performed, but |page_id| is not modified.

On the other hand, CapturePageInfo is executed as a delayed function.
This function may handle a different page if user took an action to
navigate a page, or a page is navigated automatically by a script or
a client redirect. The page can be incomplete when this function runs.
There, |page_id| can be a hint to distinguish pages and abort functions.
But if the page is redirected by refresh meta tag, it is useless.

This change omit to call CapturePageInfo if a loaded page contains
refresh meta tag to avoid TranslateHelper is confused.

BUG=236781
TEST=none

Review URL: https://chromiumcodereview.appspot.com/14769004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@201645 4ff67af0-8c30-449e-8e8b-ad334ec8d88c