Enable promotion of RenderLayers with transitions for high-DPI
This enables promotion of RenderLayers with transitions on opacity, transform,
or filter on high-DPI devices. It also provides flags that can be used to
enable this elsewhere or to disable this on high-DPI devices, and adds a
corresponding about:flags experiment.

This depends on Blink CL http://crrev.com/14651028.

BUG=178119

Review URL: https://chromiumcodereview.appspot.com/15304007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@201711 4ff67af0-8c30-449e-8e8b-ad334ec8d88c