Pressing ALT+Shift+M does not repeatedly open multiple Files.app instance.
When we launched Files.app from the application launcher, multiple
Files.app instance does not open. The Files.app item of application 
launcher stores the list of existing Files.app's windows and sets focus 
one of the windows if the item is clicked repeatedly.

This CL changed to delegate the behavior of the shortcut key to the launcher
item's method, which is called as the launcher item is clicked.

BUG=238901
TEST=Press ALT+Shift+M repeatedly.

Review URL: https://chromiumcodereview.appspot.com/14696014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@201719 4ff67af0-8c30-449e-8e8b-ad334ec8d88c