Identity API: Add component app for auth flow UI
All Identity API UI is moving into a webview tag inside a component
app, allowing Chrome Apps to be isolated from the browser.

This change adds a component app that displays a webview based dialog
box, and a private API used to launch the app and load its resources.

BUG=228908
(This is part of step #6 in the description.)

Review URL: https://chromiumcodereview.appspot.com/14768009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@201736 4ff67af0-8c30-449e-8e8b-ad334ec8d88c