Separate incognito and normal instance of Image Loader.
The image loader extension used to serve images for both: the incognito and the normal mode. To avoid persisting data in the incognito mode and leaking it to the normal mode, we should run a separate instance for the incognito mode. Since the image loader is now lazy initialized, there is no risk that it will cause high resource consumption.

TEST=There should be always one instance of the image loader in memory.
BUG=243234
R=yoshiki@chromium.org

Review URL: https://codereview.chromium.org/15701004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@201754 4ff67af0-8c30-449e-8e8b-ad334ec8d88c