First stages of the new add-limited-user flow: signing in and "Learn more"
Modify the create-profile dialog to only allow creation of a managed (aka limited) user
if the current user is signed in, and give an informative message otherwise.

Add a "Learn more" link to the checkbox for creating a limited user, and a descriptive
overlaythat appears when it is clicked.

BUG=234991
TEST=manual

Review URL: https://chromiumcodereview.appspot.com/14743018

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@201864 4ff67af0-8c30-449e-8e8b-ad334ec8d88c