Break test cases on inner asserts.
We were calling asserts withing a subroutines. In that case, asserts breaks the current subroutine but not the test case. This patch wraps calls to subroutines with ASSERT_NO_FATAL_FAILURE, which checks if it failed on any asserts, and if so, then terminates the test case.

TEST=browser_tests
BUG=235334

Review URL: https://chromiumcodereview.appspot.com/15771005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@202010 4ff67af0-8c30-449e-8e8b-ad334ec8d88c