Ensure that we honor settings like session restore, open a set of urls, when the ASH session is started.
We also need to ensure that if there are browsers open in the desktop, the ASH session on startup
behaves like the new window operation, i.e. a window with a blank tab.

We use the StartupBrowserCreatorImpl::Launch function to ensure that the session startup settings etc
are automatically honored. I had to plumb the desired host desktop type through a bunch of functions
in this class.

BUG=243072
R=sky@chromium.org
TEST=Should be covered with browser_tests when we get them running in ASH. This is tracked here http://code.google.com/p/chromium/issues/detail?id=179830

Review URL: https://codereview.chromium.org/15788002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@202258 4ff67af0-8c30-449e-8e8b-ad334ec8d88c