Don't create an unbadged shortcut when first profile is created.
This fixes all sorts of issues where the first run would create 
a new shortcut on the desktop, when it may not be desired. 

This code was originally added to handle the case where you 
delete your last profile (which would delete its shortcut), 
and then when a new profile would be created in its place, 
the desktop would no longer have the shortcut it had 
originally. 

However, in the current code deleting the last profile will 
not delete the regular Chrome shortcut off the desktop 
(when only one profile is left, the shortcut is not a profile 
specific shortcut, but a plain shortcut to the exe), since 
DeleteDesktopShortcutsAndIconFile() only looks for profile 
shortcuts. 

Thus the code this is deleting is no longer necessary to 
handle the case it was originally added for and can be 
safely removed.

Also, updates unit tests appropriately.

BUG=232940
TEST=Install Chrome with a master prefs file that says 
not to create a shortcut and launch it; a shortcut 
shouldn't be added. (See bug for more details).

Review URL: https://chromiumcodereview.appspot.com/15986006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@202293 4ff67af0-8c30-449e-8e8b-ad334ec8d88c