Fixed leak in StorageSchemaManifestHandlerTest.
This test uses the global ManifestHandlerRegistry, and its result depend on
the order of execution of other tests. If any test before created a
ManagedValueStoreCache then the ManifestHandlerRegistry has a
StorageSchemaManifestHandler, which will try to validate the bad extension
manifests, and will make this test fail.

This temporary fix always registers a StorageSchemaManifestHandler, and the
test doesn't assume that the extension is always created anymore. The long
term fix is to make ScopedTestingManifestHandlerRegistry public so that tests
like this can have a scoped ManifestHandlerRegistry.

TBR=kalman@chromium.org

Review URL: https://chromiumcodereview.appspot.com/16077003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@202461 4ff67af0-8c30-449e-8e8b-ad334ec8d88c