BrowserContext should simply own DownloadManager
This patch changes BrowserContext so it store a DownloadManager
directly (as UserData), instead of storing a refptr to it.

Client code (mostly tests) that used refptrs was changed to use either
raw pointers or to use scoped_ptr/ScopedVector in case of
DownloadManagers created by it.

The Bind() in DownloadManagerImpl::CheckForFileRemoval() is now using
a weak pointer to itself.

BUG=237871
TEST=content_unittests and {unit,browser}_tests with *Download* and *SavePage*

Review URL: https://chromiumcodereview.appspot.com/15701007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@202743 4ff67af0-8c30-449e-8e8b-ad334ec8d88c