Add metadata to content::PasswordForm to keep track of the password usage.
The motivating force behind this change is we want to ping the first time
a generated password is used. If we need to keep information about first use,
we might as well keep total usage stats as it is an interesting UMA statistic 
as well.

Changes to the other PasswordStore implementations will be in separate CLs

BUG=240560

Review URL: https://chromiumcodereview.appspot.com/14811010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@202788 4ff67af0-8c30-449e-8e8b-ad334ec8d88c