Add private API to notify end of initialization.
To be able to use input method just after log-in, we should notify the end of
ime initialization to input method engine, otherwise onActivate event will be
missed.

The right way to fix this issue is migrating packaged app and working with lazy 
background page, but it requires much effort both for chrome and extension side
so and hard to merge to M28. So please let me add this temporary hack.

BUG=242864
TEST=Manually checked on link

Review URL: https://chromiumcodereview.appspot.com/15912004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@202842 4ff67af0-8c30-449e-8e8b-ad334ec8d88c