[Downloads] Allow acquiring dangerous download file.
A dangerous download can be accepted by the user or rejected. If
rejected, the downloaded file used to be deleted. This change adds
DownloadItem::AcquireFileAndDeleteDownload() which allows the caller to
acquire the dangerous file.

The intended consumer of this feature is SafeBrowsing where the caller
may want to acquire a dangerous file that's being discarded for the
purpose of further analysis.

Also change the logic during shutdown to no longer delete dangerous
downloads, but to cancel them.

TBR=rdsmith
BUG=244604

Review URL: https://chromiumcodereview.appspot.com/14947007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@202925 4ff67af0-8c30-449e-8e8b-ad334ec8d88c