ContentSettingImageView cleanup, phase 3.
Swapping between BoxLayouts when the background toggles is confusing, and
requires us to also add and remove children in order to get the padding right.

Instead, just manually Layout(), which isn't hard, and allows us to create the
text label in the constructor and just toggle its visibility instead of actually
adding and removing it.

This CL contains a few minor changes:
* The label now obeys not only the passed-in font but also the color and
  y-position.
* The label is now ALIGN_LEFT, which probably doesn't matter with NO_ELIDE but
  is more logical in principle (since we think of the label as being pinned to
  the icon on its left side).
* The size calculations now cause the animation to smoothly grow from the size
  of the icon all the way out (and vice versa), rather than having the starting
  state of the animation be a jump (compared to the non-animating state) to
  immediately showing the background already expanded out to the margins around
  the icon.  This is almost imperceptible at full speed but makes the animation
  slightly smoother.

BUG=none
TEST=none
R=sky@chromium.org

Review URL: https://codereview.chromium.org/16115010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@203023 4ff67af0-8c30-449e-8e8b-ad334ec8d88c