Close all windows when app shim quits.
This adds a new IPC message QuitApp which is sent by the shim when the user
quits the shim (right click -> Quit). The AppShimHost responds by closing all
windows associated with the app. The shim actually quits when the extension
eventually closes and AppShimHost closes the channel to the shim.

BUG=168080
TEST=Start an app using its shim.
Right click the app in the dock and select quit.
The shell windows of that app should close.

Review URL: https://chromiumcodereview.appspot.com/14579005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@203098 4ff67af0-8c30-449e-8e8b-ad334ec8d88c