Don't open the 'stats for nerds' in a non-incognito window when incognito is forced.
Incognito mode may be always-on by configuration policy. When that is set then
the TaskManager shouldn't open a new non-incognito window.

The TaskManager doesn't have an associated Profile, so it uses the last used
Profile. This change introduces a getter for the last used allowed Profile,
which returns the OTR Profile if that's dictated by policy.

BUG=244246

Review URL: https://chromiumcodereview.appspot.com/16072014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@203128 4ff67af0-8c30-449e-8e8b-ad334ec8d88c