Auto-color-readability support for the labels inside the various location bar bubbles (EV, keyword, content setting).  The bubbles sample their background images, then compute the real background color by blending with the location bar background color, then finally pass that to the relevant Labels so they can color themselves correctly.
This also moves the actual background images to be declared in the subclasses
instead of the LocationBarView, which doesn't need to know about them.

The code in content_setting_image_view.cc is copy-and-pasted from
icon_label_bubble_view.cc.  I hope to fix this later by using the latter as a
base class for the former.

BUG=none
TEST=none
R=sky@chromium.org

Review URL: https://codereview.chromium.org/16244003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@203213 4ff67af0-8c30-449e-8e8b-ad334ec8d88c