Collect spelling service feedback on Mac
This CL enables collecting feedback for spelling service on Mac. Mac uses a
different code path for spellchecking, because Mac's built-in spellcheck lives 
in the browser process. Other operating systems keep the built-in spellcheck in
the renderer process. Other operating systems are already collecting the 
feedback for spelling service.

This CL also modifies the IPC message "SpellCheckHostMsg_RequestTextCheck" to
send a list of existing misspellings (SpellCheckMarker) together with the
spellcheck request from the renderer to the browser on Mac. Other operating
systems use the "SpellCheckHostMsg_CallSpellingService" message, which already
passes a list of existing misspellings with the spellcheck request.

BUG=170514

Review URL: https://chromiumcodereview.appspot.com/15861019

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@203227 4ff67af0-8c30-449e-8e8b-ad334ec8d88c