InstantExtended: turn off instant by default.
Adds a new finch flag ("allow_instant"). Unless this flag is present and set to
true, the Instant checkbox will not be shown and Instant will be off. When we
are ready to turn on Instant again, it will be as simple as setting
"allow_instant:1" in the finch configs.

BUG=240904

Review URL: https://chromiumcodereview.appspot.com/15115006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@203288 4ff67af0-8c30-449e-8e8b-ad334ec8d88c