Convert NetworkMenu to use NetworkStateHandler
This cleans up the separators and moves
Searching/Initializing messages to match the status area.

Additionally it removes Lock/IsLocked and related code from NetworkLibrary. This code was designed to prevent enabling / disabling a device during activation, but was only used in NetworkMenu which is only used in the Login UI which does not include activation.

BUG=167231
R=gspencer@chromium.org, pneubeck@chromium.org

Review URL: https://codereview.chromium.org/15486008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@203451 4ff67af0-8c30-449e-8e8b-ad334ec8d88c