cros: Don't wait for a page load to vet login success.
There might not be a page load to wait for. For a new user, the wait is satisfied
by NTP tab. For an old user, the wait is dismissed with a random component
extension background page or sometime it times out and causing test flakes.

BUG=241789
TEST=login_loginCryptohomeMounted no longer flakes and does not regress
     login_loginSuccess and desktopui_UrlFetch.

Review URL: https://chromiumcodereview.appspot.com/15711010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@203513 4ff67af0-8c30-449e-8e8b-ad334ec8d88c