Formatting adjustments to the "Create user" dialog.
Per UX, the bottom section is for actions, not settings/options. So the buttons
to create or remove a desktop shortcut when editing an existing user should stay
down there, but the checkboxes to select whether to create a shortcut, and
whether the user should be a limited user, need to move up into the content area.

Also per recent UX mocks, the "Name:" label should be next to its input, and
left-justified.

BUG=234991
TEST=manual

Review URL: https://chromiumcodereview.appspot.com/15813009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@203522 4ff67af0-8c30-449e-8e8b-ad334ec8d88c