Reland r203814.  Original message:
Increase omnibox size by 2 px.  Increase nominal font size in omnibox to 16 px.

The old omnibox background asset was semitransparent (for CrOS) and had to be
drawn-over with the appropriate background color.  The new assets split the
border (used by everyone) from the filling (used by CrOS alone).  As a result,
when the filling is not in use, we can now draw the background behind the
omnibox instead of in front, resulting in better antialiasing.  (Sadly, we still
need to draw a shape that fits "inside" the borders since the borders are also
semitransparent and have to overlay the toolbar in some places and the omnibox
background in others; this is unavoidable unless we want to change to a fixed-
color or programmatically-drawn omnibox border.)

All PNGs were compressed with build/sanitize-png-files.sh -o2.

BUG=231005
TEST=Omnibox is 2 px larger, contents still look OK
R=sky@chromium.org
Review URL: https://codereview.chromium.org/16286003


git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@203816 4ff67af0-8c30-449e-8e8b-ad334ec8d88c