Fixed extension/app window maximum size calculation on Windows.
Previously the height of window title bars was added to the maximum
content size of windows only if there was a maximum width as well as a
maximum height. This fixes that so it's possible to specify maxHeight
without specifying maxWidth to chrome.app.window.create().

BUG=246289

Review URL: https://chromiumcodereview.appspot.com/15969028

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@203962 4ff67af0-8c30-449e-8e8b-ad334ec8d88c