No longer trigger signin errors on connection_failed for sync.
If ProfileSyncService gets a transient network error trying to contact the
backend, we no longer treat this as a fatal signin error.

Also, we no longer display a "wait for the backend to startup" spinner in cases
where we know the spinner will immediately be dismissed, to avoid a gnarly
infinite loop in the UI code.

BUG=244769

Review URL: https://chromiumcodereview.appspot.com/16173006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@203968 4ff67af0-8c30-449e-8e8b-ad334ec8d88c