Make sure OneClickSigninSyncStarter detects Browser object being closed
Before this fix, OneClickSigninSyncStarter didn't detect when the Browser object it points to was closed. It could then cause a crash by making calls on the destroyed Browser object. OneClickSigninSyncStarter now observes the browser list to detect when the Browser it points to is closed.

BUG=235439

Review URL: https://chromiumcodereview.appspot.com/16092009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@204021 4ff67af0-8c30-449e-8e8b-ad334ec8d88c