Use DownloadItem::IsDone to check for terminal downloads.
When downloads can be resumed, a download transitioning to INTERRUPTED
state is not necessarily entering a terminal state. Therefore, a
download being in IN_PROGRESS state is not a sufficient indicator of
whether a download may eventually complete. Use DownloadItem::IsDone()
to indicate this.

TBR=estade
BUG=7648

Review URL: https://chromiumcodereview.appspot.com/16007017

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@204526 4ff67af0-8c30-449e-8e8b-ad334ec8d88c