mac: Plumbing for the notification settings view.
Since the settings view shows up in the tray, but
MessageCenter::ShowNotificationSettingsDialog() needs to be called to hook up
the provider, I changed ShowNotificationSettingsDialog() to return the delegate.
That way the tray can use that to talk to the settings controller.

No visible changes yet.

BUG=238247
R=mukai@chromium.org, rsesek@chromium.org

Review URL: https://codereview.chromium.org/16408010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@204540 4ff67af0-8c30-449e-8e8b-ad334ec8d88c