Handle key sequence for toggling voice search.  This is a no-op until ontogglevoicesearch is added server-side.
This will be used in lieu of the current server-side implementation of this keyboard shortcut to handle cases where Chrome is capturing keystrokes (see internal bug b/9123020).  In the future, we may also use this API to initiate voice search from an icon in the omnibox.

BUG=230417

Review URL: https://chromiumcodereview.appspot.com/15721006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@204677 4ff67af0-8c30-449e-8e8b-ad334ec8d88c