PNaCl component: check manifest as sanity check instead of checking llc.
Previously we would check for pnacl_public_${arch}_llc_nexe.
This lets us know if the ${arch} is correct, because it happens
to be encoded within the filename. However, we are renaming the files,
so that would not be a reliable check in the future.

Instead, read in the pnacl_public_pnacl_json file and check if the arch
specified by the manifest file is legit.

BUG=none

Review URL: https://chromiumcodereview.appspot.com/15901016

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@205024 4ff67af0-8c30-449e-8e8b-ad334ec8d88c