Display the list of retained files in the extension permissions dialog.
The dialog allows the user to revoke access for the app, restarting it
if it's currently running. If the app/extension has not retained access
to any files, the dialog is unchanged.

[Mac] XIB change:
- Enable "Show Vertical Scroller" for the scroll view.

Dialog screenshots: https://code.google.com/p/chromium/issues/detail?id=224684#c6

BUG=224684
TEST=Install an app that uses chrome.fileSystem.retainEntry and open a
file with it. In chrome://extensions/ click the permissions link for
that app. The filename should be listed.

Review URL: https://chromiumcodereview.appspot.com/15943007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@205028 4ff67af0-8c30-449e-8e8b-ad334ec8d88c