Files.app: Fixed loading the startup preference.
Originally, due to the order of initialization, the startup preference is
initialized before it is loaded. I fixed the order of initialization.

In addiiton, this CL added the 'listType' entry to startup preference that
stores which view (file list, thumbnail grid) is used at last time.

BUG=245072
TEST=manually

Review URL: https://chromiumcodereview.appspot.com/15722006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@205156 4ff67af0-8c30-449e-8e8b-ad334ec8d88c