Fix redrawing layer backed toolbars on window activation
When runing Chrome with --use-core-animation toolbars wouldn't update
their background color on window activation and deactivation.

The problem was that -[FramedBrowserWindow windowMainStatusChanged] was
simply calling setNeedsDisplay: on the frame view. This didn't
propigate to toolbars.

Fix was to explicitly listen to activate / deactivate notifications in
the toolbar code.

BUG=245900
TEST=Run Chrome with --use-core-animation. Open a new window and verify
that toolbars in the background window repaint correctly.

Review URL: https://chromiumcodereview.appspot.com/16576009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@205236 4ff67af0-8c30-449e-8e8b-ad334ec8d88c