Make Enter confirm/submit option overlays.
This change adds a 'default-button' capability to overlays. If
a button in the button strip has a 'default-button' class,
hitting enter should now activate that button, as long as the
focus is not on another button, textarea or anchor. The default
button will also be styled differently, to indicate to users
which action will be taken when hitting enter.

BUG=116859
TEST=All option dialogs/overlays should still work.

Review URL: https://chromiumcodereview.appspot.com/14361005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@205478 4ff67af0-8c30-449e-8e8b-ad334ec8d88c