Stop passing the OOBE WebUI object to AccessibilityManager
When enabling spoken feedback on an OOBE screen, ChromeVox needs to be
injected into the screen's WebUI. The AccessibilityManager expects the
caller to determine whether an OOBE screen is being shown and if so, to
pass in a pointer to its WebUI.

This has led to many different code paths trying to determine the OOBE
WebUI, all doing it slightly differently. Many code paths do not even try
at all, passing in NULL unconditionally.

This CL moves the logic that determines whether an OOBE screen is being
shown into the AccessibilityManager, removing the need to pass a WebUI
object to it.

BUG=225956
TEST=browser and unit tests updated + manual in VM

TBR=jam (testing_automation_provider_chromeos.cc)
TBR=nkostylev (tray_accessibility_browsertest.cc + core_oobe_handler.cc)
TBR=derat (chrome_shell_delegate_chromeos.cc)

Review URL: https://chromiumcodereview.appspot.com/16570002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@205516 4ff67af0-8c30-449e-8e8b-ad334ec8d88c