Re-try variations seed fetch earlier if it fails the first time.
Makes it so that if a variations seed fetch fails the first time due
to an error in the network stack (not a bad response from the server),
that we'll re-try again after 5 minutes instead of waiting for the full
fetch interval to elapse (5 hours).

Also applies to mobile platforms, where we'll also re-try in 5 minutes
in such cases, rather than on the next browser startup.

BUG=247820

Review URL: https://chromiumcodereview.appspot.com/15675019

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@205575 4ff67af0-8c30-449e-8e8b-ad334ec8d88c