Implement GetVoucherFile and GetHmonitor functions for PPB_Flash_DRM
This adds 2 functions to PPB_Flash_DRM which are used for Flash DRM-related
features. GetVoucherFile() returns a FileRef to a flash voucher file. This file
will be dropped in to the tree next to the flash DLL and will be named
plugin.vch.

GetHmonitor returns the HMONITOR (on windows only) associated with the monitor displaying the instance. cpu@ will implement the backend for this.

BUG=242241
R=cpu@chromium.org, jam@chromium.org, jschuh@chromium.org, yzshen@chromium.org

Review URL: https://codereview.chromium.org/15079003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@205640 4ff67af0-8c30-449e-8e8b-ad334ec8d88c