High DPI support for themes: Raw images for NTP attribution and background
Adding support for the NTP background and attribution theme
images. They are stored in the so-called raw images in
BrowserThemePack. Added computation of images for missing scales from
availables scales. Changed *.css and *.js files so that the URLs
contain the requisite 1x or 2x indication. (This depends on Issue 248189)

Reviewers:
pkotwicz: overall
xiyuan: for *.css and *.js

BUG=133934
R=pkotwicz@chromium.org
TEST=manual

Review URL: https://chromiumcodereview.appspot.com/16610002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@205862 4ff67af0-8c30-449e-8e8b-ad334ec8d88c