First work towards new "avatar" menu.
This is not complete but it marks a point where others can try it and build off of it for other features (and thus work in parallel).  The two main buttons are currently non-functional (to be worked on in parallel) but you can switch profiles by selecting the appropriate icon.  The "sign-out" button does work.

Example:  http://www/~bcwhite/new-avatar-2.png

To test this, add these two flags to the command line:
  --new-profile-management --gaia-profile-info

BUG=239800

Review URL: https://chromiumcodereview.appspot.com/15145006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@206092 4ff67af0-8c30-449e-8e8b-ad334ec8d88c