[sync] Force gaia re-auth on CrOs during sign in after changing password
On Chrome OS, when we detect that a user's gaia password has changed, we
prompt the user to sign out and sign in again via a link on the settings
page.

The first time the user signs out and back in, we do a local account
re-auth instead of a full-fledged gaia re-auth. As a result, we continue
to use an old OAuth2 token, and this requires the user to sign out and
sign in yet again before the token is refreshed.

This patch adds code that invalidates the existing OAuth2 token during
the first sign out, so that the user is forced to do a full-fledged
gaia-reauth during the first subsequent sign in, and thereby gets a
fresh OAuth2 token.

TBR=xiyuan@chromium.org, zelidrag@chromium.org
BUG=234164
TEST=Sign in, change gaia password, click "Sign out and sign in", and make sure you go through a gaia re-auth the first time

Review URL: https://codereview.chromium.org/16870004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@206203 4ff67af0-8c30-449e-8e8b-ad334ec8d88c