In components/autofill, move common/ to shared/
This change is part of moving autofill into its eventual structure of a layered
component. autofill/components/shared/ is a new directory that will hold code
that is shared by all platforms (including iOS). This CL populates that
directory with the code that is currently in autofill/components/common/; a
follow-up CL will move the contents of autofill/components/browser into this
directory as well.

This CL additionally imposes the correct DEPS restrictions on
autofill/components/shared/, with temporary allowances for the current
violating includes.

TBR=joi, thakis, benm

BUG=247015

Review URL: https://chromiumcodereview.appspot.com/17084005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@206593 4ff67af0-8c30-449e-8e8b-ad334ec8d88c