Pepper: Add initial browser host for FileRef.
This is the first change to introduce browser support for FileRef. It only
supports internal filesystems. This also includes the messsages that need to be
introduced (and the logic for reading and writing the new FileRef_CreateInfo
structure).

This change was tested (as part of a larger local change) and passed tests in
FileRef.

TBR=darin
BUG=225441

Review URL: https://chromiumcodereview.appspot.com/16478002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@206864 4ff67af0-8c30-449e-8e8b-ad334ec8d88c