Localize Chrome Apps directory name, and add Chrome Canary and Chromium.
This adds a ".localized" folder and a <locale>.strings file so that OSX
will display the localized folder name.
This also uses a different apps folder for Canary and Chromium.

BUG=168080
TEST=Change the preferred language to something other than English.
Restart Chrome and create a shortcut.
The Chrome Apps folder name should be localized.

Review URL: https://chromiumcodereview.appspot.com/16855005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@206945 4ff67af0-8c30-449e-8e8b-ad334ec8d88c