Get rid of RequestRegistry (part 1): remove the notion of suspended requests.
This is the first part of the series of RequestRegistry removal.
Now the class is doing nothing; it is just deleting Requests.
Everything can now be deleted, but to cleanly decouple the
dependency of other parts to it, I'll remove the class step by step.

What this patch does is:

* Remove unittest.
* Suspend => Finish.
* Resume => Start.

Previously "resume" searched existing suspended request and took over it,
but we don't need such complexity anymore.

BUG=164098

Review URL: https://chromiumcodereview.appspot.com/17066007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@206970 4ff67af0-8c30-449e-8e8b-ad334ec8d88c