Files.app: Stop applying to the fade-in animation to drag thumbnails.
Originally, the fade-in animation is applied to the drag thumbanils.  But drag
thumbnails is not real element but just a snapshot image that is taken at the
beginning of the drag.  Thus, the snapshot is taken when the fade-in effect
starts, which causes the invisible drag thumbnails.

BUG=248433
TEST=manually

Review URL: https://chromiumcodereview.appspot.com/17274004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@207240 4ff67af0-8c30-449e-8e8b-ad334ec8d88c