Resizing huge favicons when retrieving those.
 Favicons and touch icons have no reason to ever be larger than 114 pixels. As
we need to re-encode those on the browser process, very large favicon have an
impact on the CPU used on the browser process. It also means bigger messages
passed from a process to the other. This CL add a parameter so that the browser
process can specify the maximal size of the images it is interested in, so that
the renderer can resize the images if none match the criteria.

BUG=248561

Review URL: https://chromiumcodereview.appspot.com/17069002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@207303 4ff67af0-8c30-449e-8e8b-ad334ec8d88c