Move instant support to SearchTabHelper. This will also fix the search for RTL languages.
Previously, instant_tab_ never "loses" instant support. But an instant_tab_ can lose the instant support by navigating to a non-extended page.

In this CL, whenever a navigation entry is committed, we reset the instant support state of the instant_tab_ to INSTANT_SUPPORT_UNKNOWN and we determine the support once the page contents are loaded.

BUG=247727, 246323
TEST=none

Review URL: https://chromiumcodereview.appspot.com/16035020

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@207428 4ff67af0-8c30-449e-8e8b-ad334ec8d88c