Update the context menu items in the OSX App Launcher.
For old-style apps, this hides the "Open as window" checkbox menu item
from Mac (which doesn't support it). Correctly shows Open fullscreen as
fullscreen (Open maximized is not supported on Mac).

For packaged apps, it will additionally show a "Create shortcut" option
(no plural), which invokes the same behaviour as the new tab page.

BUG=138633
TEST=Launch the OSX App Launcher and observe changes above. Try creating
a shortcut for a packaged app that doesn't currently have a shortcut.

Review URL: https://chromiumcodereview.appspot.com/17371003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@207640 4ff67af0-8c30-449e-8e8b-ad334ec8d88c