When app shortcuts are enabled, do a once-off creation of all shortcuts.
This adds a boolean (kAppShortcutsEnabled) to user prefs. At startup,
ShortcutManager checks the pref and sets the current value. On a
transition from false to true, ShortcutManager iterates over all
apps and creates shortcuts as if the app is being installed.

BUG=249189
TEST=Delete app shortcuts from the OS applications folder.
Start Chrome, the shortcuts should be created.
Delete the shortcuts again.
Restart Chrome, the shortcuts are not created a second time.

Review URL: https://chromiumcodereview.appspot.com/17261016

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@207832 4ff67af0-8c30-449e-8e8b-ad334ec8d88c