Fix some timing and display issues with the supervised-user confirmation.
Collect the custodian's signed-in status each time the confirmation
dialog is shown, to catch changes in status that happen after the
Settings page loads.

Also HTML-escape the new profile's name (and the custodian's email,
for good measure), so the confirmation-dialog string can be HTML rather
than text and the link to the management page can be active.

BUG=252860, 252691
TEST=manual

Review URL: https://chromiumcodereview.appspot.com/17155020

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@208010 4ff67af0-8c30-449e-8e8b-ad334ec8d88c