[Sync] Add logic to reassociate tab nodes after restart.
Change SessionModelAssociator and TabNodePool to reuse old tab nodes.
Old tab nodes are now added to TabNodePool and if there is an
existing old tab node with syncid of tab, the tab is reassociated
with that node.
Note: In order for tab reassociation to work, the syncid of a tab
should be persisted. This CL just introduces the sync changes for
reassociating tab nodes. Persisting syncid with tab will be done
in a separate CL for Desktop, and in a downstream CL for Android.

BUG=139670,12549,139666
TEST=Includes unit tests + MANUAL testing on Android.

Review URL: https://chromiumcodereview.appspot.com/16421003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@208024 4ff67af0-8c30-449e-8e8b-ad334ec8d88c