Omnibox: Add Shortcuts as InstantExtended Provider
We might as well now that InstantExtended is using the native omnibox.  
Besides, it has a positive effect on quality.

For now, I left the framework that allows InstantExtended to use a
different set of providers.  I'll leave it up to Jered or Brett to decide
if they're ready to get rid of it.

BUG=252892

Review URL: https://chromiumcodereview.appspot.com/17563007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@208071 4ff67af0-8c30-449e-8e8b-ad334ec8d88c