Improve OSX App Launcher location for non-mouse activations and autohide docks.
This change sets the OSX app launcher level to NSDockWindowLevel, so it
shares a z-order stack with the dock and (on 10.7+) IME language input
windows. For an autohide dock, this means that it will initially appear
on top of the dock, but if the dock is hidden and reshown, the dock will
be on top.

It also checks whether the mouse cursor is visible. If it is not, assume
the app launcher was not activated using the dock. Instead, show the app
list in the bottom-right of the most likely work area. This is the
display containing the active Chrome window, if Chrome is the active
application. If Chrome is not the active application, use the primary
display, which is the one with the OSX menu bar on it.

Additionally (when the cursor is visible) and the dock is set to
autohide, move the window at least 50 pixels away from the screen edge.
This is the best we can achieve without knowing the actual size of an
autohide dock, but is enough to allow the dock to be interacted with,
e.g., to click the dock icon to hide again, and works well combined with
showing the app list on top of the dock window.

BUG=138633, 247450
TEST=Activate the app launcher on OSX with {left, bottom,
right}x{autohide, shown}x{keyboard,mouse} permutations, to ensure
requirements above are met.

Review URL: https://chromiumcodereview.appspot.com/17538002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@208454 4ff67af0-8c30-449e-8e8b-ad334ec8d88c