Move image_decoder.cc/.h from webkit\glue to content\child.
This also brings in a new content API DecodeImage which is declared in content\public\child\image_decoder_utils.h.
Used by chrome\browser code which runs mostly in the utilty process to decode icons. We still have issues with single process
mode which results in chrome\browser depending on content_child. This will be addressed in a subsequent change.

This is continuation of the ongoing work for bug https://code.google.com/p/chromium/issues/detail?can=2&q=237249
which is to split chrome.dll into a browser and a renderer/child component due to build issues on Windows. To achieve
this we need to ensure that the browser code does not depend on webkit. This mostly comes through webkit\glue.

BUG=237249
R=jam@chromium.org, jamesr@chromium.org

Review URL: https://codereview.chromium.org/17704002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@208736 4ff67af0-8c30-449e-8e8b-ad334ec8d88c