Fix bug that MouseEventPreconnect field trial was not synced to renderer process.
r203120 introduced MouseEventPreconnect field trial, but the renderer process didn't receive its data, as it was not used in the browser process. This patch adds a function SetupMouseEventPreconnectFieldTrial, which marks the field trial as used in the browser process. This will ensure the field trial data to be transferred to the renderer process.

BUG=240959

Review URL: https://chromiumcodereview.appspot.com/17749008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@208947 4ff67af0-8c30-449e-8e8b-ad334ec8d88c