The browser importer code which runs in the utility process should not depend on chrome\browser data types like TemplateURL, etc.
This CL removes the usage of chrome\browser\search_engines\template_url** from the ExternalProcessImporterBridge class.
To achieve this, a structure URLKeywordInfo has been added to the importer_data_types.h file. The
ImporterBridge::SetKeywords virtual function has been changed to accept a vector of the URLKeywordInfo structures.
This is converted to TempateURL instances in the InProcessImporterBridge class which lives in the browser process.

The other change is in the Firefox3 importer where in we pass the Firefox profile XML data in a vector to the browser
via the newly added virtual function ImporterBridge::SetFirefoxSearchEnginesXMLData. The ExternalProcessImporterBridge 
implementation of this function sends over the XML data via an IPC to the browser. The InProcessImporterBridge implementation
parses this data to retrieve the search engine information which is then imported.

This is continuation of the ongoing work for bug https://code.google.com/p/chromium/issues/detail?can=2&q=237249
which is to split chrome.dll into a browser and a renderer/child component due to build issues on Windows. To achieve
this we need to ensure that the browser code does not depend on webkit/child. 

BUG=237249
R=gab@chromium.org, jschuh@chromium.org, scottmg@chromium.org

Review URL: https://codereview.chromium.org/18064002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@209072 4ff67af0-8c30-449e-8e8b-ad334ec8d88c