chromeos: PowerManagerHandler -> PowerStatus; move to ash.
With this change, ash::internal::PowerStatus is mostly the
same as the old chromeos::PowerManagerHandler: it's a
singleton that asks chromeos::PowerManagerClient to request
the power status at startup and then provides
PowerSupplyStatus structs to various classes in ash/system.

Going forward, I plan to remove the PowerSupplyStatus struct
(so that the PowerStatus class hangs on to the original
PowerSupplyProperties protocol buffers sent by the power
manager) and add methods to PowerStatus for accessing the
interesting parts of the protobufs (instead of having
multiple copies of the status sanitized and stored across
the different classes in ash/system).

BUG=254173
TBR=sky@chromium.org

Review URL: https://chromiumcodereview.appspot.com/18023007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@209103 4ff67af0-8c30-449e-8e8b-ad334ec8d88c