cros: Record cpu/paging metrics at some input events
This creates several histograms in the form

Platform.SwapJank.<event>.<measure>.<swap group>.<time interval>

where

<event> is TabSwitch or Scroll
<measure> is Cpu or Faults
<swap group> is Swap0, ... Swap3
<time interval> is Time1, ... Time4

When an event occurs (tab switch or scroll), we post tasks to the blocking
pool to read some measures at a few time intervals following the event
(100, 300, 1000 and 3000 milliseconds).  Each measure is an average
for the interval.  Measures are CPU utilization (%) and number of page
faults per second.  We further classify the event by the current
utilization of the swap device: up to 10%, 10-30%, 30-60%, >60%.

The purpose of these histograms is to characterize the impact of
memory pressure (and consequently paging) on common events.  We chose
tab switches and scroll events because they seem to be the ones
that are most affected by paging.  Ideally we would like to measure
the time from the input event to the completion of the request,
but this is hard to do, so instead we attempt to extrapolate the impact
of paging from these indirect measures.

This has been tested both on Linux and Chrome OS, and the histograms
appear reasonable.

BUG=252772
TEST=manual, run with --v=0 --vmodule=swap_metrics=1 and switch tabs
R=semenzato@chromium.org, stevenjb@chromium.org

Review URL: https://codereview.chromium.org/17577003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@209225 4ff67af0-8c30-449e-8e8b-ad334ec8d88c