Don't close window when adding users.
The enter keyevent was double-handled in the 'Manage other users'
options dialog, making the dialog close when a user was added. This
CL fixes the problem by changing the event handled from keypress to
keydown so it matches the other event, then calling stopPropagation
on it when it is handled. 

BUG=254163

Review URL: https://chromiumcodereview.appspot.com/17733004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@209253 4ff67af0-8c30-449e-8e8b-ad334ec8d88c