Add UMA metrics to measure the effectiveness of views fuzzing
Add new buckets to the Ash.GestureTarget histogram which
will be useful in measuring the effectiveness of views 
fuzzing when it is implemented. The idea is to count taps 
made on the different element types located on/near the 
tab strip. A description of the new types added to
GestureActionType in touch_uma.cc:

GESTURE_TABSWITCH_TAP: A tap on a currently unselected
tab (i.e., the tap was responsible for selecting the tab).

GESTURE_TABNOSWITCH_TAP: A tap on a currently selected
tab (i.e, the tap had no visible effect).

GESTURE_TABCLOSE_TAP: A tap on a tab's close button.

GESTURE_NEWTAB_TAP: A tap on the new tab button.

GESTURE_FRAMEMAXIMIZE_TAP: A tap on the frame maximize
button.

GESTURE_FRAMEVIEW_TAP: A tap on the space to the right of
the tabstrip, left of the tabstrip, or in between tabs.
Such a tap has no visible effect.

GESTURE_ROOTVIEWTOP_TAP: A tap on the top edge of a browser
window. Such a tap has no visible effect.

GESTURE_MAXIMIZE_DOUBLETAP: A double-tap on the browser
window resulting in a maximize or restore action.

BUG=248230

Review URL: https://chromiumcodereview.appspot.com/16832003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@209263 4ff67af0-8c30-449e-8e8b-ad334ec8d88c