Add separate PowerIdleAction prefs for AC and battery power
This CL replaces the PowerIdleAction pref with a PowerAcIdleAction pref
and a PowerBatteryIdleAction pref, allowing different idle actions to be
specified when running on AC and battery power. This is a prerequisite
for allowing Chrome to control power management on the login screen where
different actions are needed.

Enterprise user policies are update accordingly, replacing IdleAction with
IdleActionAC and IdleActionBattery.

BUG=254404
TEST=Updated browser and unit tests

Review URL: https://chromiumcodereview.appspot.com/18051012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@209265 4ff67af0-8c30-449e-8e8b-ad334ec8d88c