[Mac][Panels] Make it easier to minimize panels.
This enables window minimization of panels through the standard Cmd+M shortcut.
If the panel is detached, it goes into the dock like a normal window, but if
it's attached, then it will be minimized to the bottom of the screen.

This also enables minimization behavior when the titlebar is double-clicked.

BUG=104394,255139,177821
TEST=Double clicking on a panel titlebar minimizes it. Cmd+M does, too.

Review URL: https://chromiumcodereview.appspot.com/18177010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@209343 4ff67af0-8c30-449e-8e8b-ad334ec8d88c