Stop allowing passing empty file path to DownloadFileRequest.
Currently it allows to pass empty path as the target of downloading,
and then it automatically creates a temporary file and returns
the path. However, it seems that the default is not used at all.

Let's remove it and change the configuration interface to the
same style as the other options.

BUG=none

Review URL: https://chromiumcodereview.appspot.com/18312004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@209437 4ff67af0-8c30-449e-8e8b-ad334ec8d88c