Fix and reenable PanelBrowserTest.AnimateBounds
We should create a detached panel, not docked panel for the test. This is because docked panel cannot be moved to any random position.

Also reenable the test for Linux. Though there is no animation on Linux, we can still run part of the test.

BUG=179069
TEST=test reenabled
R=petewil@chromium.org

Review URL: https://codereview.chromium.org/18223007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@209447 4ff67af0-8c30-449e-8e8b-ad334ec8d88c