Add DeleteRecursively to AsyncFileUtil.
On Drive file system, recursive deletion can be done more efficiently than
removing each file/directory. To support it, this CL adds an optional method
DeleteRecursively to AsyncFileUtil, and use it (if possible) in
LocalFileSystemOperation.
This is the preparation of merging {Remote,Local}FileSystemOperation. Currently
the new method is not used on any platforms, but will be used by Drive File
System.

BUG=110121
TEST=Ran unit_tests

Review URL: https://chromiumcodereview.appspot.com/18314002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@209484 4ff67af0-8c30-449e-8e8b-ad334ec8d88c