Add an explicit way of forcing the MDnsClient to listen on the network
Add an object called MDnsListenRef, whose existence forces the MDnsClient to
continue listening on the network, but who has no other function.

Rationale: Since a sandboxed process using an MDnsClient may not be able
to create new sockets on demand, it makes sense to force the MDnsClient to
be awake for the lifetime of the process. Add the ability to force the
MDnsClient to stay awake.

Also useful for abstractions over MDnsClient where you know that it makes sense
not to immediately shut down the client even if we shut down all listeners and
transactions.

BUG=233821

Review URL: https://chromiumcodereview.appspot.com/17922002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@209812 4ff67af0-8c30-449e-8e8b-ad334ec8d88c