Correct RTL alignment of page title on chrome://policy
Currently the page header is aligned too far to the left, when viewing
the chrome://policy page in a right-to-left language. This looks
especially bad when scrolling down (see screenshots attached to bug).
This change aligns the page title to the right hand side in RTL mode.

BUG=253860

Review URL: https://chromiumcodereview.appspot.com/17587016

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@210167 4ff67af0-8c30-449e-8e8b-ad334ec8d88c