Allow Chrome OS login profile to have different default pref values
This CL is a prerequisite for putting Chrome in charge of all power
management settings. Since the power management defaults on the login
screen differ from those during a user session, the default values of
the power prefs should be set differently for the login profile.

This CL has no functional changes. It just does the following:

1/ Split chrome::RegisterUserPrefs() into:
 * chrome::RegisterUserProfilePrefs()
 * chrome::RegisterLoginProfilePrefs()

2/ Rename all other RegisterUserPrefs() methods to RegisterProfilePrefs().

BUG=241794
TEST=Updated unit and browser tests

TBR=jochen@chromium.org (blanket rename of RegisterUserPrefs() in chrome/)
TBR=asargent@chromium.org (blanket rename of RegisterUserPrefs() in apps/)
TBR=joi@chromium.org (blanket rename of RegisterUserPrefs() in components/)

Review URL: https://chromiumcodereview.appspot.com/18199003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@210310 4ff67af0-8c30-449e-8e8b-ad334ec8d88c