Fixing destruction order in login database unit tests
Before, the ScopedTempDir would be destroyed before
LoginDatabase, which would still be locking the database file
within the temporary folder. This would cause an error in
ScopedTempDir's d-tor, and the temporary folder would stay
on disk.

Now the order of LoginDatabaseTest's members is aligned to
the order of initialization in SetUp and LoginDatabase is
destroyed before ScopedTempDir. This also makes the tests
run a lot faster on my Windows machine (>4s to 110ms).

TEST=unit_tests.exe --gtest_filter="LoginDatabaseTest.*"
BUG=

Review URL: https://chromiumcodereview.appspot.com/18684002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@210374 4ff67af0-8c30-449e-8e8b-ad334ec8d88c