Android: Speech Recognition infobar and tray notification.
Refactor ChromeSpeechRecognitionManagerDelegate in order to reuse it on
Android: Extract the x-webkit-speech input bubble UI (unused on Android)
into a child class, and use the child class only if ENABLE_INPUT_SPEECH
is defined (not the case on Android). In effect, the delegate makes a
MediaAccessRequest, prompting the user for permission to record audio
with an infobar, and indicating when recording is in progress with help
of a sticky android status tray notification.

BUG=243367

Review URL: https://chromiumcodereview.appspot.com/17261030

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@210414 4ff67af0-8c30-449e-8e8b-ad334ec8d88c