drive: Merge "Clean local cache" and "Reload"
The behaviors of the two actions are similar and it's hard to see the difference
from the menu text strings alone. Merge the two behaviors into "Clean local cache"
and remove "Reload". What we need is one reset button that can be used as a last
resort when things are messed up.

BUG=256952
TEST=Open Drive in Files.app; Open the gear menu with Ctrl-pressed. Confirm that 'Reload' is gone and 'Clear local cache' works (remove cached files and reload Drive)
R=asargent@chromium.org, yoshiki@chromium.org

Review URL: https://codereview.chromium.org/18673003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@210491 4ff67af0-8c30-449e-8e8b-ad334ec8d88c