Update the HTML Media Capture implementation.
In the latest version of the spec, the `capture' attribute is now a boolean
which simply indicates whether the device's capabilities should be used or
not (choosing between cameras, microphones etc is up to the user agent based
on the contents of the `accept' attribute).

This particular patch updates the type of the `capture' member of
content::FileChooserParams by passing an std::pair consisting of a vector of
MIME types passed to the `accept' attribute of the <input> tag and a boolean
corresponding to the `capture' attribute.

The behavior in Android has been slightly changed due to how the spec
change. Namely:
 o Sites that specify only the "capture" attribute setting the "accept"
   attribute will now get a generic picker.
 o Sites that specify the "capture" attribute and pass "*/*" to "accept"
   will now get a generic picker.
 o Sites that specify the "capture" and multiple MIME types for the "accept"
   attribute will now get a generic picker.

R=abarth@chromium.org,beverloo@chromium.org,bulach@chromium.org,tedchoc@chromium.org,eseidel@chromium.org
BUG=240252

Review URL: https://chromiumcodereview.appspot.com/14758008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@210553 4ff67af0-8c30-449e-8e8b-ad334ec8d88c