Rename RenderViewGone IPC/methods to better reflect reality.
The RenderViewGone IPC and methods are incorrectly named, as those are used when a renderer process has died. Since we are passing in termination status around and usually iterating over all Render{Widget|View}Host in a specific process, it actually applies to the whole process. This CL renames all applicable instances of this method and the IPC message associated with it.

BUG=258087

Review URL: https://chromiumcodereview.appspot.com/18339006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@210727 4ff67af0-8c30-449e-8e8b-ad334ec8d88c