Implement getting the monitor (HMONITOR) for flash
So we properly track the monitor associated with the native widget associated with renderwidget that is associated
with the renderview that is assoicated with the pp_instance.

Please note that the test loops since it can take a bit for the task executes in the UI thread.

See bug for more details.

BUG=249135
TEST=see bug

Review URL: https://chromiumcodereview.appspot.com/18438004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@211046 4ff67af0-8c30-449e-8e8b-ad334ec8d88c