Added SigninManager.java and signin_manager_android.cc.
This class handles the common sign-in flow on the Java side, and now checks
if the account has policy enabled. If so then policy is fetched and enforced
before starting signed-in services such as sync, which may be disabled by the
policy.

The sign-in and sign-out handling on Android has been moved out of
ProfileSyncServiceAndroid and into the new class.

The policy behavior is currently disabled, since Android builds don't have
ENABLE_CONFIGURATION_POLICY by default.

TBR=atwilson@chromium.org
BUG=255887

Review URL: https://chromiumcodereview.appspot.com/18281003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@211059 4ff67af0-8c30-449e-8e8b-ad334ec8d88c